Drop requirement to register meta[name] extensions
The original spirit behind the requirement to register `meta[name]` extensions
was good but in practice, it has ended up costing a huge amount of maintenance
time for something that's produced almost no real benefit to end users.

Most values registered at https://wiki.whatwg.org/wiki/MetaExtensions are
essentially for private use for only the people/organizations who minted them,
and have zero general utility and no public applications that consume them.

So the HTML checker quit emitting errors for unregistered `meta[name]` values 10
months ago and there are no plans to re-add that reporting to it ever again. So
this is a case where the spec requirement is essentially a fiction, since the
main intended implementor of the requirement is wilfully choosing to ignore it.